backend/requirements.txt
fastapi
uvicorn
boto3
mangum
python-multipart
anthropic
openai
httpx
requests
beautifulsoup4
lxml
PyJWT[crypto]
